+++ src/main/java/com/balsamic/sejongmalsami/object/AdminDto.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.springframework.data.domain.Page;
 
 @Builder
 @Getter
@@ -14,6 +15,7 @@ import lombok.ToString;
 @ToString
 public class AdminDto {
   private Member member;
+  private Page<Member> membersPage;
   private Yeopjeon yeopjeon;
   private YeopjeonHistory yeopjeonHistory;
 }
