+++ src/main/java/com/balsamic/sejongmalsami/service/MemberService.java
@@ -417,32 +417,4 @@ public class MemberService {
         .build();
   }
 
-  // 관리자-회원관리 : 필터링 검색
-  public MemberDto getFilteredMembers(MemberCommand command) {
-    return MemberDto.builder()
-        .membersPage(
-            memberRepository.findAllDynamic(
-                command.getStudentId(),
-                command.getStudentName(),
-                command.getUuidNickname(),
-                command.getMajor(),
-                command.getAcademicYear(),
-                command.getEnrollmentStatus(),
-                command.getAccountStatus(),
-                command.getRole(),
-                command.getLastLoginStart(),
-                command.getLastLoginEnd(),
-                command.getIsFirstLogin(),
-                command.getIsEdited(),
-                command.getIsDeleted(),
-                PageRequest.of(
-                    command.getPageNumber(),
-                    command.getPageSize(),
-                    Sort.by(Sort.Direction.fromString(command.getSortDirection()),
-                        command.getSortField())
-                )
-            )
-        )
-        .build();
-  }
 }
